Refactor user tests, add API tests for user
